fix_4. Score is 100

+++ tasks/tutorial/aplusb/aplusb.hpp
@@ -1,5 +1,5 @@
 #include <iostream>
 
-long long Sum(int a, int b) {
+int64_t Sum(int a, int b) {
     return a + b;
 }
